IoT/requirements.txt
Flask
requests>=2.7.0
opencv-python
numpy
Pillow